+++ tencentcloud/provider.go
@@ -66,6 +66,7 @@ Data Sources
   tencentcloud_route_table
   tencentcloud_security_group
   tencentcloud_security_groups
+  tencentcloud_ssl_certificates
   tencentcloud_subnet
   tencentcloud_vpc
   tencentcloud_vpc_instances
@@ -151,6 +152,9 @@ MySQL Resources
 Redis Resources
   tencentcloud_redis_instance
   tencentcloud_redis_backup_config
+
+SSL Resources
+  tencentcloud_ssl_certificate
 
 VPC Resources
   tencentcloud_vpc
@@ -256,6 +260,7 @@ func Provider() *schema.Provider {
 			"tencentcloud_gaap_security_policies":      dataSourceTencentCloudGaapSecurityPolices(),
 			"tencentcloud_gaap_security_rules":         dataSourceTencentCloudGaapSecurityRules(),
 			"tencentcloud_gaap_certificates":           dataSourceTencentCloudGaapCertificates(),
+			"tencentcloud_ssl_certificates":            dataSourceTencentCloudSslCertificates(),
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
@@ -320,6 +325,7 @@ func Provider() *schema.Provider {
 			"tencentcloud_gaap_certificate":           resourceTencentCloudGaapCertificate(),
 			"tencentcloud_gaap_security_policy":       resourceTencentCloudGaapSecurityPolicy(),
 			"tencentcloud_gaap_security_rule":         resourceTencentCloudGaapSecurityRule(),
+			"tencentcloud_ssl_certificate":            resourceTencentCloudSslCertificate(),
 		},
 
 		ConfigureFunc: providerConfigure,